+++ samples/xform/src/matrix.h
@@ -1,6 +1,9 @@
 #ifndef MATRIX_H
 #define MATRIX_H
 
-void mat_mul_3x3_3x1(float s1[3][3], float* s2, float* dest);
+#include "fixed_maths.h"
+#include <stdint.h>
+
+void mat_mul_3x3_3x1(int32_t s1[3][3], int32_t* s2, int32_t* dest);
 
 #endif
